Replaced big decimal with double everywhere

+++ src/main/java/lv/proofit/busapp/api/tax/rates/TaxRatesResponse.java
@@ -4,11 +4,9 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
 import lombok.Value;
 
-import java.math.BigDecimal;
-
 @Value
 @Builder
 public class TaxRatesResponse {
     @Schema(example = "21")
-    BigDecimal taxRate;
+    double taxRate;
 }
