+++ src/main/java/lv/proofit/busapp/feature/draft/price/BasePriceService.java
@@ -3,15 +3,13 @@ package lv.proofit.busapp.feature.draft.price;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-import java.math.BigDecimal;
-
 @Service
 @RequiredArgsConstructor
 public class BasePriceService {
 
     private final BasePriceRepository repository;
 
-    public BigDecimal getBy(String busTerminalName) {
+    public double getBy(String busTerminalName) {
         return repository.findByTerminalName(busTerminalName)
                 .orElseThrow(() -> new IllegalArgumentException("Could not find bus terminal with name:["+ busTerminalName +"]"))
                 .getPrice();
